fix: fix bug update photo profile

+++ app/Services/UserService.php
@@ -14,7 +14,7 @@ class UserService
         if ($request->hasFile('photo')) {
             $data['photo'] = $request->file('photo')->store('photo');
 
-            if (Storage::exists(auth()->user()->photo)) {
+            if (Storage::exists(auth()->user()->photo) && !is_null(auth()->user()->photo)) {
                 Storage::delete(auth()->user()->photo);
             }
         }
